First class ECS schema support
Object mappers now have a schema parameters.
At the moment, the only valid value is `ecs`.
During dynamic mapping, just before date/numeric detection and dynamic templates are evaluated,
it is evaluated whether the field name is defined by the schema.
If it is, the mapping is determined by the built-in schema file.

I've used a modified version of the ecs schema generation to store a schema file with flat mappings.